+++ ECommerce.ProductApi/ProductApi.Domain/Constants/DBConstraint.cs
@@ -5,6 +5,6 @@ public static class DBConstraint
     public const int PRODUCT_NAME_MAX_LENGTH = 250;
     public const int PRODUCT_MIN_QUANTITY = 1;
     public const int PRODUCT_MAX_QUANTITY = 10000;
-    public const int PRODUCT_MIN_PRICE = 1;
-    public const int PRODUCT_MAX_PRICE = 1000000;
+    public const decimal PRODUCT_MIN_PRICE = 1.00m;
+    public const decimal PRODUCT_MAX_PRICE = 1000000.00m;
 }
